interpolate matrices instead of euler angles
interpolation occured on 3 translation values and 3 euler angles values,
which lead to numerical singularities problems.

I now occurs on the full transform matrix (16 values). As the
interpolation method is STEP, we should avoid the problem of
non-homeogeneous interpolateds matrices but the sample point
should be chosen near enough.

Note: the test is not updated yet.